+++ core/chains/evm/config/chain_specific_config.go
@@ -128,8 +128,8 @@ func setChainSpecificConfigDefaultSets() {
 		gasEstimatorMode:                      "BlockHistory",
 		gasFeeCapDefault:                      *DefaultGasFeeCap,
 		gasLimitDefault:                       DefaultGasLimit,
+		gasLimitMax:                           DefaultGasLimit, // equal since no effect other than Arbitrum
 		gasLimitMultiplier:                    1.0,
-		gasLimitMax:                           DefaultGasLimit * 100,
 		gasLimitTransfer:                      21000,
 		gasPriceDefault:                       *DefaultGasPrice,
 		gasTipCapDefault:                      *DefaultGasTip,
@@ -289,10 +289,11 @@ func setChainSpecificConfigDefaultSets() {
 	arbitrumMainnet.chainType = config.ChainArbitrum
 	arbitrumMainnet.gasBumpThreshold = 0 // Disable gas bumping on arbitrum
 	arbitrumMainnet.gasEstimatorMode = "Arbitrum"
-	arbitrumMainnet.gasLimitDefault = 10_000_000
-	arbitrumMainnet.gasLimitMax = 500_000_000
-	arbitrumMainnet.gasLimitTransfer = 80_000
-	arbitrumMainnet.minGasPriceWei = *big.NewInt(0)           // Arbitrum uses the suggested gas price so we don't want to place any limits on the minimum
+	arbitrumMainnet.gasLimitMax = 1_000_000_000
+	arbitrumMainnet.minGasPriceWei = *big.NewInt(0)          // Arbitrum uses the suggested gas price so we don't want to place any limits on the minimum
+	arbitrumMainnet.gasPriceDefault = *big.NewInt(100000000) // 0.1 gwei
+	arbitrumMainnet.maxGasPriceWei = *assets.GWei(1000)
+	arbitrumMainnet.gasFeeCapDefault = *assets.GWei(1000)
 	arbitrumMainnet.blockHistoryEstimatorBlockHistorySize = 0 // Force an error if someone set GAS_UPDATER_ENABLED=true by accident; we never want to run the block history estimator on arbitrum
 	arbitrumMainnet.linkContractAddress = "0xf97f4df75117a78c1A5a0DBb814Af92458539FB4"
 	arbitrumMainnet.ocrContractConfirmations = 1