+++ core/chains/evm/gas/arbitrum_estimator.go
@@ -22,6 +22,7 @@ type ArbConfig interface {
 	EvmGasLimitMax() uint32
 }
 
+//go:generate mockery --name ethClient --output ./mocks/ --case=underscore --structname ETHClient
 type ethClient interface {
 	CallContract(ctx context.Context, msg ethereum.CallMsg, blockNumber *big.Int) ([]byte, error)
 }
@@ -83,6 +84,10 @@ func (a *arbitrumEstimator) Close() error {
 }
 
 // GetLegacyGas estimates both the gas price and the gas limit.
+//   - Price is delegated to the embedded l2SuggestedPriceEstimator.
+//   - Limit is computed from the dynamic values perL2Tx and perL1CalldataUnit, provided by the getPricesInArbGas() method
+//     of the precompilie contract at ArbGasInfoAddress. perL2Tx is a constant amount of gas, and perL1CalldataUnit is
+//     multiplied by the length of the tx calldata. The sum of these two values plus the original l2GasLimit is returned.
 func (a *arbitrumEstimator) GetLegacyGas(calldata []byte, l2GasLimit uint32, maxGasPriceWei *big.Int, opts ...Opt) (gasPrice *big.Int, chainSpecificGasLimit uint32, err error) {
 	gasPrice, _, err = a.Estimator.GetLegacyGas(calldata, l2GasLimit, maxGasPriceWei, opts...)
 	if err != nil {
@@ -105,7 +110,6 @@ func (a *arbitrumEstimator) GetLegacyGas(calldata []byte, l2GasLimit uint32, max
 			}
 		}
 		perL2Tx, perL1CalldataUnit := a.getPricesInArbGas()
-		// TODO is it ok to return lsGasLimit if unitialized?
 		chainSpecificGasLimit = l2GasLimit + perL2Tx + uint32(len(calldata))*perL1CalldataUnit
 		a.logger.Debugw("GetLegacyGas", "l2GasLimit", l2GasLimit, "calldataLen", len(calldata), "perL2Tx", perL2Tx,
 			"perL1CalldataUnit", perL1CalldataUnit, "chainSpecificGasLimit", chainSpecificGasLimit)
@@ -149,51 +153,72 @@ func (a *arbitrumEstimator) run() {
 	}
 }
 
-// refreshPricesInArbGas calls getPricesInArbGas() on the precompile contract 0x000000000000000000000000000000000000006c.
+// refreshPricesInArbGas calls getPricesInArbGas() and caches the refreshed prices.
 func (a *arbitrumEstimator) refreshPricesInArbGas() (t *time.Timer) {
 	t = time.NewTimer(utils.WithJitter(a.pollPeriod))
 
-	ctx, cancel := evmclient.ContextWithDefaultTimeoutFromChan(a.chStop)
-	defer cancel()
-
-	// @return (per L2 tx, per L1 calldata unit, per storage allocation)
-	// function getPricesInArbGas() external view returns (uint256, uint256, uint256);
-	//
-	// https://github.com/OffchainLabs/nitro/blob/f7645453cfc77bf3e3644ea1ac031eff629df325/contracts/src/precompiles/ArbGasInfo.sol#L69
-	precompile := common.HexToAddress("0x000000000000000000000000000000000000006c")
-	b, err := a.client.CallContract(ctx, ethereum.CallMsg{
-		To:   &precompile,
-		Data: common.Hex2Bytes("02199f34"),
-	}, big.NewInt(-1))
+	perL2Tx, perL1CalldataUnit, err := a.callGetPricesInArbGas()
 	if err != nil {
-		a.logger.Warnf("Failed to refresh prices, got error: %s", err)
+		a.logger.Warnw("Failed to refresh prices", "err", err)
 		return
 	}
 
+	a.logger.Debugw("refreshPricesInArbGas", "perL2Tx", perL2Tx, "perL2CalldataUnit", perL1CalldataUnit)
+
+	a.getPricesInArbGasMu.Lock()
+	a.perL2Tx = perL2Tx
+	a.perL1CalldataUnit = perL1CalldataUnit
+	a.getPricesInArbGasMu.Unlock()
+	return
+}
+
+const (
+	// ArbGasInfoAddress is the address of the "Precompiled contract that exists in every Arbitrum chain."
+	// https://github.com/OffchainLabs/nitro/blob/f7645453cfc77bf3e3644ea1ac031eff629df325/contracts/src/precompiles/ArbGasInfo.sol
+	ArbGasInfoAddress = "0x000000000000000000000000000000000000006C"
+	// ArbGasInfo_getPricesInArbGas is the a hex encoded call to:
+	// `function getPricesInArbGas() external view returns (uint256, uint256, uint256);`
+	ArbGasInfo_getPricesInArbGas = "02199f34"
+)
+
+// callGetPricesInArbGas calls ArbGasInfo.getPricesInArbGas() on the precompile contract ArbGasInfoAddress.
+//
+// @return (per L2 tx, per L1 calldata unit, per storage allocation)
+// function getPricesInArbGas() external view returns (uint256, uint256, uint256);
+//
+// https://github.com/OffchainLabs/nitro/blob/f7645453cfc77bf3e3644ea1ac031eff629df325/contracts/src/precompiles/ArbGasInfo.sol#L69
+func (a *arbitrumEstimator) callGetPricesInArbGas() (perL2Tx uint32, perL1CalldataUnit uint32, err error) {
+	ctx, cancel := evmclient.ContextWithDefaultTimeoutFromChan(a.chStop)
+	defer cancel()
+
+	precompile := common.HexToAddress(ArbGasInfoAddress)
+	b, err := a.client.CallContract(ctx, ethereum.CallMsg{
+		To:   &precompile,
+		Data: common.Hex2Bytes(ArbGasInfo_getPricesInArbGas),
+	}, big.NewInt(-1))
+	if err != nil {
+		return 0, 0, err
+	}
+
 	if len(b) != 3*32 { // returns (uint256, uint256, uint256);
-		a.logger.Errorf("Failed to refresh prices, return data length (%d) different than expected (%d)", len(b), 3*32)
+		err = fmt.Errorf("return data length (%d) different than expected (%d)", len(b), 3*32)
 		return
 	}
 	bPerL2Tx := new(big.Int).SetBytes(b[:32])
 	bPerL1CalldataUnit := new(big.Int).SetBytes(b[32:64])
 	// ignore perStorageAllocation
 	if !bPerL2Tx.IsUint64() || !bPerL1CalldataUnit.IsUint64() {
-		a.logger.Errorf("Failed to refresh prices, returned integers are not uint64", "perL2Tx", bPerL2Tx.String(),
-			"perL1CalldataUnit", bPerL1CalldataUnit.String())
+		err = fmt.Errorf("returned integers are not uint64 (%s, %s)", bPerL2Tx.String(), bPerL1CalldataUnit.String())
+		return
 	}
 
-	perL2Tx := bPerL2Tx.Uint64()
-	perL1CalldataUnit := bPerL1CalldataUnit.Uint64()
-	if perL2Tx > math.MaxUint32 || perL1CalldataUnit > math.MaxUint32 {
-		a.logger.Errorf("Failed to refresh prices, returned integers are not uint32", "perL2Tx", perL2Tx,
-			"perL1CalldataUnit", perL1CalldataUnit)
+	perL2TxU64 := bPerL2Tx.Uint64()
+	perL1CalldataUnitU64 := bPerL1CalldataUnit.Uint64()
+	if perL2TxU64 > math.MaxUint32 || perL1CalldataUnitU64 > math.MaxUint32 {
+		err = fmt.Errorf("returned integers are not uint32 (%d, %d)", perL2TxU64, perL1CalldataUnitU64)
+		return
 	}
-
-	a.logger.Debugw("refreshPricesInArbGas", "perL2Tx", perL2Tx, "perL2CalldataUnit", perL1CalldataUnit)
-
-	a.getPricesInArbGasMu.Lock()
-	defer a.getPricesInArbGasMu.Unlock()
-	a.perL2Tx = uint32(perL2Tx)
-	a.perL1CalldataUnit = uint32(perL1CalldataUnit)
+	perL2Tx = uint32(perL2TxU64)
+	perL1CalldataUnit = uint32(perL1CalldataUnitU64)
 	return
 }