+++ core/chains/evm/gas/cmd/arbgas/main.go
@@ -1,3 +1,4 @@
+// arbgas takes a single URL argument and prints the result of three GetLegacyGas calls to the Arbitrum gas estimator.
 package main
 
 import (
@@ -26,10 +27,10 @@ func main() {
 	defer func() { _ = sync() }()
 	lggr.SetLogLevel(zapcore.DebugLevel)
 
-	run(context.Background(), lggr, url, func(e gas.Estimator) {
+	withEstimator(context.Background(), lggr, url, func(e gas.Estimator) {
 		printGetLegacyGas(e, make([]byte, 10), 500_000, assets.GWei(1))
 		printGetLegacyGas(e, make([]byte, 10), 500_000, assets.GWei(1), gas.OptForceRefetch)
-		printGetLegacyGas(e, make([]byte, 10), 5_000_000, assets.GWei(1))
+		printGetLegacyGas(e, make([]byte, 10), max, assets.GWei(1))
 	})
 }
 
@@ -43,13 +44,15 @@ func printGetLegacyGas(e gas.Estimator, calldata []byte, l2GasLimit uint32, maxG
 	fmt.Println("Limit:", limit)
 }
 
-func run(ctx context.Context, lggr logger.Logger, url string, f func(e gas.Estimator)) {
+const max = 50_000_000
+
+func withEstimator(ctx context.Context, lggr logger.Logger, url string, f func(e gas.Estimator)) {
 	rc, err := rpc.Dial(url)
 	if err != nil {
 		log.Fatal(err)
 	}
 	ec := ethclient.NewClient(rc)
-	e := gas.NewArbitrumEstimator(lggr, &config{max: 50_000_000}, rc, ec)
+	e := gas.NewArbitrumEstimator(lggr, &config{max: max}, rc, ec)
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
 	err = e.Start(ctx)
@@ -61,6 +64,8 @@ func run(ctx context.Context, lggr logger.Logger, url string, f func(e gas.Estim
 	f(e)
 }
 
+var _ gas.ArbConfig = &config{}
+
 type config struct {
 	max uint32
 }
@@ -68,5 +73,3 @@ type config struct {
 func (c *config) EvmGasLimitMax() uint32 {
 	return c.max
 }
-
-var _ gas.ArbConfig = &config{}